January/22th-Jan-Thurs/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "passlib[cryptography]>=1.7.4",
    "pathlib[bcrypt]>=1.0.1",
    "pymysql>=1.1.2",
    "python-decouple>=3.8",
    "python-jose[crytography]>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
